Streamlinging player joins
Made it so that the player joining the game is a separate call to the
server. Instead of joining immediately on connecting, we are ensuring
that user input is valid before joining the game.

If the input is not valid, then highlight the input fields that are
invalid

Otherwise, inform the client of the # of players and if there are at
least 2 players, let them start the game

Once the game is started, each player gets the information private to
them at the start of the game.